requirements.txt
# ===============================
# Core DL
# ===============================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ===============================
# Medical Imaging
# ===============================
monai>=1.3.0
nibabel>=5.1.0
SimpleITK>=2.3.0

# ===============================
# Multi-GPU / Training
# ===============================
accelerate>=0.25.0
tensorboard>=2.14.0

# ===============================
# Config / Utils
# ===============================
pyyaml>=6.0
easydict>=1.10
tqdm>=4.66.0

# ===============================
# Scientific
# ===============================
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.22.0
scikit-learn>=1.3.0

# ===============================
# Optional but recommended
# ===============================
matplotlib>=3.8.0
pandas>=2.1.0
